Rename validation task document model
