+++ emmet-core/emmet/core/vasp/material.py
@@ -11,7 +11,7 @@ from emmet.core.material import MaterialsDoc as CoreMaterialsDoc
 from emmet.core.material import PropertyOrigin
 from emmet.core.structure import StructureMetadata
 from emmet.core.vasp.calc_types import CalcType, RunType, TaskType
-from emmet.core.vasp.task import TaskDocument
+from emmet.core.vasp.task_valid import TaskDocument
 
 
 SETTINGS = EmmetSettings()
@@ -171,7 +171,8 @@ class MaterialsDoc(CoreMaterialsDoc, StructureMetadata):
 
     @classmethod
     def construct_deprecated_material(
-        cls, task_group: List[TaskDocument],
+        cls,
+        task_group: List[TaskDocument],
     ) -> "MaterialsDoc":
         """
         Converts a group of tasks into a deprecated material
