+++ emmet-core/emmet/core/vasp/validation.py
@@ -10,7 +10,7 @@ from emmet.core.settings import EmmetSettings
 from emmet.core.base import EmmetBaseModel
 from emmet.core.mpid import MPID
 from emmet.core.utils import DocEnum
-from emmet.core.vasp.task import TaskDocument
+from emmet.core.vasp.task_valid import TaskDocument
 
 SETTINGS = EmmetSettings()
