+++ tests/emmet-core/vasp/test_materials.py
@@ -5,7 +5,7 @@ from monty.io import zopen
 
 from emmet.core.vasp.calc_types import TaskType
 from emmet.core.vasp.material import MaterialsDoc
-from emmet.core.vasp.task import TaskDocument
+from emmet.core.vasp.task_valid import TaskDocument
 
 
 @pytest.fixture
